img_study_buddy/img_study_buddy/requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.2
django-crispy-forms==2.0
django-formtools==2.4.1
django-quill-editor==0.1.40
django-tinymce==3.6.1
packaging==23.1
patsy==0.5.3
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
